Backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
langchain-google-genai
langchain